Add a version number to API URL
Normally I very much dislike this approach, because it forces you to completely copy all of your API views once you decide to bump the version number. However in this case, we only have a very small number of API endpoints, and a client base that might possibly be outdated without our control, so it makes sense.