2017.1.0
The first release where I feel like, yes, this is something I can show publicly. Lot's of stuff is missing still, but the core of ablator is there, and it's a clear indication of where this is going.

Version numbering for now will be `<Year>.<Major>.<Minor>`, which should be compatible with semantic versioning.

The mascot of this release is a panda! 🐼