Tag-based Rollout Strategies (#126)
Right now, the core.functionality.which function goes through a set of pipeline functions to determine which flavor to return. Data is passed around using a context object, WhichContext, which updates during the pipeline as new data is added to it.

Into this environment enter Tags. For each tag, Managers of an App should be able to set their own rollout strategy. This association is ordered, so the rollout strategies of Tags with higher priority always precede those of Tags with lower priority.

Related: #122, which deals with the user interface of managing tags and rollout schemes.

* Add docstring for WhichContext

* Rename Release to RolloutStrategy

* Model updates

* Add priority field to RolloutStrategy

* Show Rollout Strategies in UI

* Updated URL to say Rollout Strategy instead of Release

* UI for creating and editing Tags

* Show user count in Tag Detail

* Update Wording on Client User info page

* Introduce error checking when adding or editing Rollout Strategies

* Begin to update availability pipeline to tag system

* Fix a bug that occurred with editing Functionality Names

* Create new Availabilities based on Tags/Rollout Strategies

* Update create_dummy_data management command

* Fix flake8 issues

* Fix a type checking issue

* Removing tests since they need to be completely rewritten